+++ okitweb/static/okit/js/okit_console.js
@@ -4,8 +4,8 @@
 */
 console.info('Loaded Console Javascript');
 
-const okitVersion = '0.10.2';
-const okitReleaseDate = '27th August 2020';
+const okitVersion = '0.11.0';
+const okitReleaseDate = '16th September 2020';
 // Validation
 const validate_error_colour = "#ff4d4d";
 const validate_warning_colour = "#ffd633";