+++ okitweb/static/okit/model/js/artefacts/virtual_cloud_network.js
@@ -39,7 +39,7 @@ class VirtualCloudNetwork extends OkitArtifact {
     ** Delete Processing
      */
     deleteChildren() {
-        console.groupCollapsed('Deleting Children of ' + this.getArtifactReference() + ' : ' + this.display_name);
+        console.log('Deleting Children of ' + this.getArtifactReference() + ' : ' + this.display_name);
         // Remove Subnets
         this.getOkitJson().subnets = this.getOkitJson().subnets.filter(function(child) {
             if (child.vcn_id === this.id) {
@@ -103,7 +103,16 @@ class VirtualCloudNetwork extends OkitArtifact {
             }
             return true;
         }, this);
-        console.groupEnd();
+        // Network Security Groups
+        this.getOkitJson().network_security_groups = this.getOkitJson().network_security_groups.filter(function(child) {
+            if (child.vcn_id === this.id) {
+                console.info('Deleting ' + child.display_name);
+                child.delete();
+                return false; // So the filter removes the element
+            }
+            return true;
+        }, this);
+        console.log();
     }
 
     /*