+++ okitweb/static/okit/view/designer/js/artefacts/dynamic_routing_gateway.js
@@ -12,10 +12,11 @@ class DynamicRoutingGatewayView extends OkitDesignerArtefactView {
         super(artefact, json_view);
     }
 
-    get parent_id() {return this.artefact.vcn_id;}
+    get parent_id() {return this.artefact.vcn_id && this.artefact.vcn_id !== '' ? this.artefact.vcn_id : this.artefact.compartment_id;}
+    get parent() {return this.artefact.vcn_id && this.artefact.vcn_id !== '' ? this.getJsonView().getVirtualCloudNetwork(this.parent_id) : this.getJsonView().getCompartment(this.parent_id);}
 
     getParent() {
-        return this.getJsonView().getVirtualCloudNetwork(this.parent_id);
+        return this.parent;
     }
 
     getParentId() {
@@ -26,7 +27,7 @@ class DynamicRoutingGatewayView extends OkitDesignerArtefactView {
      ** SVG Processing
      */
     draw() {
-        console.group('Drawing ' + this.getArtifactReference() + ' : ' + this.id + ' [' + this.parent_id + ']');
+        console.log('Drawing ' + this.getArtifactReference() + ' : ' + this.id + ' [' + this.parent_id + ']');
         let svg = super.draw();
         // Get Inner Rect to attach Connectors
         let rect = svg.select("rect[id='" + safeId(this.id) + "']");
@@ -49,12 +50,12 @@ class DynamicRoutingGatewayView extends OkitDesignerArtefactView {
             // Draw Connectors
             // this.drawConnectors();
         }
-        console.groupEnd();
+        console.log();
         return svg;
     }
 
     drawConnectors() {
-        console.group('Drawing Connectors for ' + this.getArtifactReference() + ' : ' + this.id + ' [' + this.parent_id + ']');
+        console.log('Drawing Connectors for ' + this.getArtifactReference() + ' : ' + this.id + ' [' + this.parent_id + ']');
         // Get Grand Parent
         let grandparent_id = d3.select(d3Id(this.parent_id)).attr('data-parent-id');
         // Define Connector Parent
@@ -93,7 +94,7 @@ class DynamicRoutingGatewayView extends OkitDesignerArtefactView {
                 }
             }
         }
-        console.groupEnd();
+        console.log();
     }
 
     // Return Artifact Specific Definition.